foundry_evm_fuzz

Module invariant

Modules

call_override 🔒

filters 🔒

Structs

ArtifactFilters

Contains which contracts are to be targeted or excluded on an invariant test through their artifact identifiers.

FuzzRunIdentifiedContracts

Contracts identified as targets during a fuzz run.

InvariantContract

Test contract which is testing its invariants.

RandomCallGenerator

Given a TestRunner and a strategy, it generates calls. Used inside the Fuzzer inspector to override external calls to test for potential reentrancy vulnerabilities..

SenderFilters

Filter for acceptable senders to use for invariant testing. Exclusion takes priority if clashing.

TargetedContract

A contract identified as target for invariant testing.

TargetedContracts

A collection of contracts identified as targets for invariant testing.