Module foundry_evm_fuzz::strategies::invariants

Functions

fuzz_contract_with_calldata

Given a function, it returns a proptest strategy which generates valid abi-encoded calldata for that function’s input types.

generate_call 🔒

Strategy to generate a transaction where the sender, target and calldata are all generated through specific strategies.

invariant_strat

Creates the invariant strategy.

override_call_strat

Given a target address, we generate random calldata.

select_random_function 🔒

Strategy to select a random mutable function from the abi.

select_random_sender 🔒

Strategy to select a sender address: