foundry_evm_fuzz::strategies::mutators

Trait AbiMutator

pub(crate) trait AbiMutator:
    Sized
    + Copy
    + Debug {
    // Required methods
    fn flip_random_bit(
        self,
        size: usize,
        test_runner: &mut TestRunner,
    ) -> Option<Self>;
    fn mutate_interesting_byte(
        self,
        size: usize,
        test_runner: &mut TestRunner,
    ) -> Option<Self>;
    fn mutate_interesting_word(
        self,
        size: usize,
        test_runner: &mut TestRunner,
    ) -> Option<Self>;
    fn mutate_interesting_dword(
        self,
        size: usize,
        test_runner: &mut TestRunner,
    ) -> Option<Self>;
}

ABI mutator that changes current value by flipping a random bit and randomly injecting interesting words - see https://github.com/AFLplusplus/LibAFL/blob/90cb9a2919faf386e0678870e52784070cdac4b6/crates/libafl/src/mutators/mutations.rs#L88-L123. Implemented for uint, int, address and fixed bytes.

Required Methods

fn flip_random_bit( self, size: usize, test_runner: &mut TestRunner, ) -> Option<Self>

fn mutate_interesting_byte( self, size: usize, test_runner: &mut TestRunner, ) -> Option<Self>

fn mutate_interesting_word( self, size: usize, test_runner: &mut TestRunner, ) -> Option<Self>

fn mutate_interesting_dword( self, size: usize, test_runner: &mut TestRunner, ) -> Option<Self>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.