Module foundry_evm_fuzz::strategies::param

Constants

MAX_ARRAY_LEN 🔒

The max length of arrays we fuzz for is 256.

Functions

fuzz_param

Given a parameter type, returns a strategy for generating values for that type.

fuzz_param_from_state

Given a parameter type, returns a strategy for generating values for that type, given some EVM fuzz state.

fuzz_param_inner 🔒

fuzz_param_with_fixtures

Given a parameter type and configured fixtures for param name, returns a strategy for generating values for that type.