Module foundry_evm_fuzz::strategies::state

Structs

FuzzDictionary

Constants

PUSH_BYTE_ANALYSIS_LIMIT 🔒

The maximum number of bytes we will look at in bytecodes to find push bytes (24 KiB).

Functions

build_initial_state

Builds the initial EvmFuzzState from a database.

collect_created_contracts

Collects all created contracts from a StateChangeset which haven’t been discovered yet. Stores them at targeted_contracts and created_contracts.

collect_push_bytes 🔒

Collects all push bytes from the given bytecode.

collect_state_from_call

Collects state changes from a [StateChangeset] and logs into an EvmFuzzState according to the given FuzzDictionaryConfig.

fuzz_calldata_from_state

Given a function and some state, it returns a strategy which generated valid calldata for the given function’s input types, based on state taken from the EVM.

Type Aliases

EvmFuzzState

A set of arbitrary 32 byte data from the VM used to generate values for the strategy.