Module foundry_evm_fuzz::strategies

Modules

calldata 🔒

int 🔒

invariants 🔒

param 🔒

state 🔒

uint 🔒

Structs

EvmFuzzState

A set of arbitrary 32 byte data from the VM used to generate values for the strategy.

IntStrategy

Value tree for signed ints (up to int256). The strategy combines 3 different strategies, each assigned a specific weight:

UintStrategy

Value tree for unsigned ints (up to uint256). The strategy combines 3 different strategies, each assigned a specific weight:

Functions

fuzz_calldata

Given a function, it returns a strategy which generates valid calldata for that function’s input types, following declared test fixtures.

fuzz_calldata_from_state

Given a function and some state, it returns a strategy which generated valid calldata for the given function’s input types, based on state taken from the EVM.

fuzz_contract_with_calldata

Given a function, it returns a proptest strategy which generates valid abi-encoded calldata for that function’s input types.

fuzz_param

Given a parameter type, returns a strategy for generating values for that type.

fuzz_param_from_state

Given a parameter type, returns a strategy for generating values for that type, given some EVM fuzz state.

fuzz_param_with_fixtures

Given a parameter type and configured fixtures for param name, returns a strategy for generating values for that type. Fixtures can be currently generated for uint, int, address, bytes and string types and are defined for parameter name.

invariant_strat

Creates the invariant strategy.

override_call_strat

Given a target address, we generate random calldata.