foundry_evm_fuzz::strategies

Module mutators

Macros

impl_increment_decrement_mutator 🔒

Statics

INTERESTING_8 🔒

INTERESTING_16 🔒

Interesting 16-bit values to inject.

INTERESTING_32 🔒

Interesting 32-bit values to inject.

Traits

AbiMutator 🔒

ABI mutator that changes current value by flipping a random bit and randomly injecting interesting words - see https://github.com/AFLplusplus/LibAFL/blob/90cb9a2919faf386e0678870e52784070cdac4b6/crates/libafl/src/mutators/mutations.rs#L88-L123. Implemented for uint, int, address and fixed bytes.

IncrementDecrementMutator 🔒

Mutator that randomly increments or decrements an uint or int.

Functions

flip_random_bit_in_slice 🔒

Flips a random bit in the given mutable byte slice.

mutate_interesting_byte_slice 🔒

Mutates a random byte in the given byte slice by replacing it with a randomly chosen interesting 8-bit value.

mutate_interesting_dword_slice 🔒

Mutates a random 4-byte (32-bit) region in the byte slice with a randomly chosen interesting 32-bit value.

mutate_interesting_word_slice 🔒

Mutates a random 2-byte (16-bit) region in the byte slice with a randomly chosen interesting 16-bit value.

validate_int_mutation 🔒

Returns mutated int value if different from the original value and if it fits in the given size, otherwise None.

validate_uint_mutation 🔒

Returns mutated uint value if different from the original value and if it fits in the given size, otherwise None.